Audio: Remove Kernel Warning in Audio Code
[Defect / Enhancement / New Feature]
MobC00284798

[Problem]
Kernel warning in Audio file during kernel build

[Solution]
Remove kernel warning by type casting CSL_CAPH_MIXER_e to int in
switch condition in sound/caph_hawaii/csl/csl_caph_srcmixer.c file.

[Reviewers]
MPS Audio Reviewers

[Testing]
    [Test Cases]
	Playback, capture, Voicecall, FM playback

    [Label/CL# Synced To]
	Head sdb-common-android-jb-4.2

    [Platform]
	HawaiiStone edn010

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/13648
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ sound/caph_hawaii/csl/csl_caph_srcmixer.c
@@ -906,7 +906,7 @@ static UInt8 csl_caph_srcmixer_get_chaloutchnl(CSL_CAPH_MIXER_e
 					       outChnl)
 {
 	UInt8 chalOutChnl = CAPH_M_NONE;
-	switch (outChnl) {
+	switch ((int)outChnl) {
 	case CSL_CAPH_SRCM_STEREO_CH1_L:
 		chalOutChnl = CAPH_M0_Left;
 		break;
